pyproject.toml
[project]
name = "utube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "edge-tts>=6.1.0",
    "moviepy>=1.0.3",
    "openai>=1.0.0",
    "openai-whisper>=20231117",
    "python-dotenv>=1.0.0",
]
